src/security.py
```python
from web3 import Web3
from web3.middleware import geth_poa_middleware
from eth_account import Account
from src.config import INFURA_URL, PRIVATE_KEY

web3 = Web3(Web3.HTTPProvider(INFURA_URL))
web3.middleware_onion.inject(geth_poa_middleware, layer=0)

def secure_transaction(transaction):
    """
    Function to secure transactions and smart contracts against common attacks.
    """
    # Check for reentrancy attacks
    check_reentrancy(transaction)

    # Check for overflow/underflow attacks
    check_overflow_underflow(transaction)

    # Sign and send the transaction
    signed_txn = web3.eth.account.sign_transaction(transaction, PRIVATE_KEY)
    return web3.eth.sendRawTransaction(signed_txn.rawTransaction)

def check_reentrancy(transaction):
    """
    Function to check for reentrancy attacks.
    """
    # TODO: Implement reentrancy check logic
    pass

def check_overflow_underflow(transaction):
    """
    Function to check for overflow/underflow attacks.
    """
    # TODO: Implement overflow/underflow check logic
    pass

def anonymize_transaction(transaction):
    """
    Function to anonymize transactions, protecting the user's identity and transaction details.
    """
    # TODO: Implement transaction anonymization logic
    pass
```