tests/test_security.py
```python
import unittest
from unittest.mock import patch
from src import security

class TestSecurity(unittest.TestCase):

    @patch('src.security.secure_transaction')
    def test_secure_transaction(self, mock_secure_transaction):
        # Test if the function is called
        mock_secure_transaction.return_value = True
        result = security.secure_transaction('transaction_data')
        self.assertTrue(result)
        mock_secure_transaction.assert_called_once_with('transaction_data')

    @patch('src.security.check_reentrancy_attack')
    def test_check_reentrancy_attack(self, mock_check_reentrancy_attack):
        # Test if the function is called
        mock_check_reentrancy_attack.return_value = False
        result = security.check_reentrancy_attack('transaction_data')
        self.assertFalse(result)
        mock_check_reentrancy_attack.assert_called_once_with('transaction_data')

    @patch('src.security.check_overflow_underflow')
    def test_check_overflow_underflow(self, mock_check_overflow_underflow):
        # Test if the function is called
        mock_check_overflow_underflow.return_value = False
        result = security.check_overflow_underflow('transaction_data')
        self.assertFalse(result)
        mock_check_overflow_underflow.assert_called_once_with('transaction_data')

    @patch('src.security.anonymize_transaction')
    def test_anonymize_transaction(self, mock_anonymize_transaction):
        # Test if the function is called
        mock_anonymize_transaction.return_value = 'anonymized_transaction_data'
        result = security.anonymize_transaction('transaction_data')
        self.assertEqual(result, 'anonymized_transaction_data')
        mock_anonymize_transaction.assert_called_once_with('transaction_data')

if __name__ == '__main__':
    unittest.main()
```